+++ RELEASES.md
@@ -25,6 +25,7 @@ Upcoming
 - Add attribute hideRegex to pipes (next to global property)
 - Fix UUIDGeneratorPipe when stubbed remove attribute type
 - Add httpRequest and httpResponse sessionkeys to WebServiceListener
+- Add facility to put outputValidator in inputValidator (to save memory)